requirements.txt
torch==2.7.0
transformers>=4.45.1,<4.47.0
librosa==0.11.0
torchvision==0.22
torchaudio==2.7.0
grpcio==1.76.0
protobuf==6.33.0
grpcio-tools==1.76.0
julius
argbind
numpy
soundfile
pyloudnorm
importlib-resources
scipy
torch
ffmpy
ipython
rich
matplotlib
pystoi
torch_stoi
flatten-dict
markdown2
randomname
tensorboard
tqdm
einops
numba
vector_quantize_pytorch
omegaconf
debugpy